correcting mistakes

+++ cleanrl/sac_continuous_action.py
@@ -352,8 +352,9 @@ poetry run pip install "stable_baselines3==2.0.0a1"
                 min_qf_next_target = torch.min(qf1_next_target, qf2_next_target) - alpha * next_state_log_pi
                 next_q_value = data.rewards.flatten() + (1 - data.dones.flatten()) * args.gamma * (min_qf_next_target).view(-1)
 
-            qf1_a_values = qf1(data.observations, data.actions, risk).view(-1)
-            qf2_a_values = qf2(data.observations, data.actions, risk).view(-1)
+            qf1_a_values = qf1(data.observations.float(), data.actions.float(), risk).view(-1)
+            qf2_a_values = qf2(data.observations.float(), data.actions.float(), risk).view(-1)
+
             qf1_loss = F.mse_loss(qf1_a_values, next_q_value)
             qf2_loss = F.mse_loss(qf2_a_values, next_q_value)
             qf_loss = qf1_loss + qf2_loss
